fix entry point on sketch
