+++ examples/basic-svg/package.json
@@ -16,7 +16,7 @@
   "author": "Jon Gold <jon.gold@airbnb.com>",
   "license": "MIT",
   "devDependencies": {
-    "@skpm/builder": "^0.4.0"
+    "@skpm/builder": "^0.7.5"
   },
   "dependencies": {
     "prop-types": "^15.5.8",
