--- src/index.common.ts
@@ -1,6 +1,0 @@
-export { Platform } from './Platform';
-export { StyleSheet } from './stylesheet';
-export { getSymbolComponentByName, getSymbolMasterByName, injectSymbols } from './symbol';
-export { useWindowDimensions } from './context';
-
-export * from './components';
